# server/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "Duration" : {
        "type" : "string",
        "format" : "duration",
        "examples" : [ "P1D" ]
      },
      "HttpProblem" : {
        "description" : "HTTP Problem Response according to RFC9457 and RFC7807",
        "additionalProperties" : true,
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "A optional URI reference that identifies the problem type",
            "examples" : [ "https://example.com/errors/not-found" ]
          },
          "title" : {
            "type" : "string",
            "description" : "A optional, short, human-readable summary of the problem type",
            "examples" : [ "Not Found" ]
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "The HTTP status code for this occurrence of the problem",
            "examples" : [ 404 ]
          },
          "detail" : {
            "type" : "string",
            "description" : "A optional human-readable explanation specific to this occurrence of the problem",
            "examples" : [ "Record not found" ]
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "description" : "A URI reference that identifies the specific occurrence of the problem",
            "examples" : [ "https://api.example.com/errors/123" ]
          }
        }
      },
      "HttpValidationProblem" : {
        "description" : "HTTP Validation Problem Response according to RFC9457 and RFC7807",
        "additionalProperties" : true,
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "A optional URI reference that identifies the problem type",
            "examples" : [ "https://example.com/errors/not-found" ]
          },
          "title" : {
            "type" : "string",
            "description" : "A optional, short, human-readable summary of the problem type",
            "examples" : [ "Not Found" ]
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "The HTTP status code for this occurrence of the problem",
            "examples" : [ 404 ]
          },
          "detail" : {
            "type" : "string",
            "description" : "A optional human-readable explanation specific to this occurrence of the problem",
            "examples" : [ "Record not found" ]
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "description" : "A URI reference that identifies the specific occurrence of the problem",
            "examples" : [ "https://api.example.com/errors/123" ]
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Violation"
            },
            "description" : "List of validation constraint violations that occurred"
          }
        }
      },
      "OffsetDateTime" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T12:15:50-04:00" ]
      },
      "PickupDropOffType" : {
        "type" : "string",
        "enum" : [ "REGULAR", "NONE", "PHONE_AGENCY", "COORDINATE_WITH_DRIVER" ]
      },
      "RailJourney" : {
        "type" : "object",
        "properties" : {
          "origin" : {
            "type" : "string"
          },
          "destination" : {
            "type" : "string"
          },
          "departureTime" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "arrivalTime" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "duration" : {
            "$ref" : "#/components/schemas/Duration"
          },
          "changes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "legs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RailJourneyLeg"
            }
          }
        }
      },
      "RailJourneyLeg" : {
        "type" : "object"
      },
      "SimpleJourney" : {
        "type" : "object",
        "properties" : {
          "departureTime" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "arrivalTime" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "duration" : {
            "$ref" : "#/components/schemas/Duration"
          },
          "changes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "legs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SimpleLeg"
            }
          }
        }
      },
      "SimpleLeg" : {
        "type" : "object",
        "properties" : {
          "origin" : {
            "type" : "string"
          },
          "destination" : {
            "type" : "string"
          },
          "departureTime" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "arrivalTime" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "originTrainUid" : {
            "type" : "string"
          },
          "destinationTrainUid" : {
            "type" : "string"
          },
          "originPickUpType" : {
            "$ref" : "#/components/schemas/PickupDropOffType"
          },
          "destinationDropOffType" : {
            "$ref" : "#/components/schemas/PickupDropOffType"
          },
          "operator" : {
            "type" : "string"
          },
          "mode" : {
            "type" : "string"
          },
          "duration" : {
            "$ref" : "#/components/schemas/Duration"
          },
          "boardingInterchange" : {
            "$ref" : "#/components/schemas/Duration"
          }
        }
      },
      "Violation" : {
        "type" : "object",
        "description" : "Validation constraint violation details",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "The field for which the validation failed",
            "examples" : [ "#/profile/email" ]
          },
          "in" : {
            "type" : "string",
            "description" : "Part of the http request where the validation error occurred such as query, path, header, form, body",
            "examples" : [ "query", "path", "header", "form", "body" ]
          },
          "message" : {
            "type" : "string",
            "description" : "Description of the validation error",
            "examples" : [ "Invalid email format" ]
          }
        }
      }
    }
  },
  "tags" : [ {
    "name" : "Journey planning",
    "description" : "RAPTOR searches over the loaded GTFS feed."
  } ],
  "paths" : {
    "/" : {
      "get" : {
        "summary" : "Pareto-optimal journeys in a time window",
        "description" : "Every journey that is not beaten on both departure and arrival, as a flat list of legs. The search covers a single service day.",
        "tags" : [ "Journey planning" ],
        "parameters" : [ {
          "description" : "Destination stop id (CRS code, e.g. `VIC`). An unknown stop yields an empty list, not an error.",
          "name" : "dest",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "required" : true
        }, {
          "description" : "End of the search window, in the same form as `startDate`. May fall on the following day, which includes GTFS after-midnight (&gt; 24:00) departures rather than dropping them.",
          "name" : "endDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "required" : true
        }, {
          "description" : "Stop ids to avoid: journeys calling at or riding through any of them are excluded. Repeat the parameter for several stops.",
          "name" : "notVia",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "default" : ""
          }
        }, {
          "description" : "Origin stop id (CRS code, e.g. `BTN`). An unknown stop yields an empty list, not an error.",
          "name" : "orig",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "required" : true
        }, {
          "description" : "Start of the search window: ISO 8601 with a UTC offset, e.g. `2026-06-03T08:30:00+01:00` or `2026-06-03T07:30:00Z`. Read on the Europe/London clock, which is the one GTFS times run on and which picks the service date. A zone-less value (`2026-06-03T08:30:00`) is still accepted and read as London wall-clock, but is deprecated and logged at WARN.",
          "name" : "startDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Matching journeys, empty if none run.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SimpleJourney"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "A window parameter is not an ISO 8601 date-time, or `orig`/`dest` also appears in `notVia`.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpProblem"
                }
              }
            }
          },
          "503" : {
            "description" : "Server at capacity: the bulkhead is full or the query exceeded its timeout. Retry after the interval in the `Retry-After` header.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpProblem"
                }
              }
            }
          }
        }
      }
    },
    "/detail" : {
      "get" : {
        "summary" : "As the range query, with full calling points",
        "description" : "The same journeys as `/`, expanded with each leg's intermediate calls, train identity and operator.",
        "tags" : [ "Journey planning" ],
        "parameters" : [ {
          "description" : "Destination stop id (CRS code, e.g. `VIC`). An unknown stop yields an empty list, not an error.",
          "name" : "dest",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "required" : true
        }, {
          "description" : "End of the search window, in the same form as `startDate`. May fall on the following day, which includes GTFS after-midnight (&gt; 24:00) departures rather than dropping them.",
          "name" : "endDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "required" : true
        }, {
          "description" : "Stop ids to avoid: journeys calling at or riding through any of them are excluded. Repeat the parameter for several stops.",
          "name" : "notVia",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "default" : ""
          }
        }, {
          "description" : "Origin stop id (CRS code, e.g. `BTN`). An unknown stop yields an empty list, not an error.",
          "name" : "orig",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "required" : true
        }, {
          "description" : "Start of the search window: ISO 8601 with a UTC offset, e.g. `2026-06-03T08:30:00+01:00` or `2026-06-03T07:30:00Z`. Read on the Europe/London clock, which is the one GTFS times run on and which picks the service date. A zone-less value (`2026-06-03T08:30:00`) is still accepted and read as London wall-clock, but is deprecated and logged at WARN.",
          "name" : "startDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Matching journeys, empty if none run.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RailJourney"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "A window parameter is not an ISO 8601 date-time, or `orig`/`dest` also appears in `notVia`.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpProblem"
                }
              }
            }
          },
          "503" : {
            "description" : "Server at capacity: the bulkhead is full or the query exceeded its timeout. Retry after the interval in the `Retry-After` header.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpProblem"
                }
              }
            }
          }
        }
      }
    },
    "/first-arrival" : {
      "get" : {
        "summary" : "Earliest arrival from a single departure time",
        "description" : "One departure time rather than a window: the journeys reaching the destination soonest when leaving at or after `startDate`.",
        "tags" : [ "Journey planning" ],
        "parameters" : [ {
          "description" : "Destination stop id (CRS code, e.g. `VIC`). An unknown stop yields an empty list, not an error.",
          "name" : "dest",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "required" : true
        }, {
          "description" : "Stop ids to avoid: journeys calling at or riding through any of them are excluded. Repeat the parameter for several stops.",
          "name" : "notVia",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "default" : ""
          }
        }, {
          "description" : "Origin stop id (CRS code, e.g. `BTN`). An unknown stop yields an empty list, not an error.",
          "name" : "orig",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "required" : true
        }, {
          "description" : "Start of the search window: ISO 8601 with a UTC offset, e.g. `2026-06-03T08:30:00+01:00` or `2026-06-03T07:30:00Z`. Read on the Europe/London clock, which is the one GTFS times run on and which picks the service date. A zone-less value (`2026-06-03T08:30:00`) is still accepted and read as London wall-clock, but is deprecated and logged at WARN.",
          "name" : "startDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OffsetDateTime"
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Matching journeys, empty if none run.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RailJourney"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "A window parameter is not an ISO 8601 date-time, or `orig`/`dest` also appears in `notVia`.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpProblem"
                }
              }
            }
          },
          "503" : {
            "description" : "Server at capacity: the bulkhead is full or the query exceeded its timeout. Retry after the interval in the `Retry-After` header.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpProblem"
                }
              }
            }
          }
        }
      }
    }
  },
  "info" : {
    "title" : "velociRAPTOR",
    "version" : "0.2.0-SNAPSHOT",
    "description" : "RAPTOR journey planning over a GTFS feed. Times on the wire are ISO 8601 with a UTC offset; the server works on the Europe/London clock, which is the one GTFS times and service dates are expressed on."
  },
  "servers" : [ {
    "url" : "/"
  } ]
}